+++ tests/server/static_endpoints_test.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"mock-server/internal/configs"
 	"mock-server/internal/control"
-	hlp "mock-server/internal/test_helpers"
 	"testing"
 )
 
@@ -24,7 +23,7 @@ func TestStaticRoutesSimple(t *testing.T) {
 	testUrl := endpoint + "/test_url"
 
 	// no routes created -> 400
-	code, body := hlp.DoGet(testUrl, t)
+	code, body := DoGet(testUrl, t)
 	if code != 400 {
 		t.Errorf("expected 400 on mismatch get")
 	}
@@ -34,7 +33,7 @@ func TestStaticRoutesSimple(t *testing.T) {
 	}
 
 	// expects []
-	code, body = hlp.DoGet(staticApiEndpoint, t)
+	code, body = DoGet(staticApiEndpoint, t)
 	if code != 200 {
 		t.Errorf("expected 200 code response on list all request")
 	}
@@ -48,13 +47,13 @@ func TestStaticRoutesSimple(t *testing.T) {
 		"path": "/test_url",
 		"expected_response": "hello"
 	}`)
-	code = hlp.DoPost(staticApiEndpoint, requestBody, t)
+	code = DoPost(staticApiEndpoint, requestBody, t)
 	if code != 200 {
 		t.Errorf("create route failed")
 	}
 
 	// expects `hello`
-	code, body = hlp.DoGet(testUrl, t)
+	code, body = DoGet(testUrl, t)
 	if code != 200 {
 		t.Errorf("expected to be possible make request to new route")
 	}
@@ -64,7 +63,7 @@ func TestStaticRoutesSimple(t *testing.T) {
 	}
 
 	// expects ["/test_url"]
-	code, body = hlp.DoGet(staticApiEndpoint, t)
+	code, body = DoGet(staticApiEndpoint, t)
 	if code != 200 {
 		t.Errorf("expected 200 code response on list all request")
 	}
@@ -78,13 +77,13 @@ func TestStaticRoutesSimple(t *testing.T) {
 		"path": "/test_url",
 		"expected_response": "hehe"
 	}`)
-	code = hlp.DoPut(staticApiEndpoint, otherRequestBody, t)
+	code = DoPut(staticApiEndpoint, otherRequestBody, t)
 	if code != 204 {
 		t.Errorf("update route failed")
 	}
 
 	// expects `hehe`
-	code, body = hlp.DoGet(testUrl, t)
+	code, body = DoGet(testUrl, t)
 	if code != 200 {
 		t.Errorf("expected to be possible make request to updated route")
 	}
@@ -94,13 +93,13 @@ func TestStaticRoutesSimple(t *testing.T) {
 	}
 
 	// detele /test_url
-	code = hlp.DoDelete(staticApiEndpoint+"?path=/test_url", t)
+	code = DoDelete(staticApiEndpoint+"?path=/test_url", t)
 	if code != 204 {
 		t.Errorf("it must be possible to delete route")
 	}
 
 	// /test_url deleted -> 400
-	code, body = hlp.DoGet(testUrl, t)
+	code, body = DoGet(testUrl, t)
 	if code != 400 {
 		t.Errorf("expected to be impossible to request deleted route: %d != 400", code)
 	}
@@ -110,7 +109,7 @@ func TestStaticRoutesSimple(t *testing.T) {
 	}
 
 	// expects []
-	code, body = hlp.DoGet(staticApiEndpoint, t)
+	code, body = DoGet(staticApiEndpoint, t)
 	if code != 200 {
 		t.Errorf("expected 200 code response on list all request")
 	}
@@ -134,12 +133,12 @@ func TestStaticRoutesDoublePost(t *testing.T) {
 		"path": "/test_url",
 		"expected_response": "hello"
 	}`)
-	code := hlp.DoPost(staticApiEndpoint, requestBody, t)
+	code := DoPost(staticApiEndpoint, requestBody, t)
 	if code != 200 {
 		t.Errorf("create route failed: expected 200 != %d", code)
 	}
 
-	code = hlp.DoPost(staticApiEndpoint, requestBody, t)
+	code = DoPost(staticApiEndpoint, requestBody, t)
 	if code != 409 {
 		t.Errorf("expected to receive conflict: expected 409 != %d", code)
 	}
@@ -148,7 +147,7 @@ func TestStaticRoutesDoublePost(t *testing.T) {
 		"path": "/test_url",
 		"expected_response": "hello"
 	}`)
-	code = hlp.DoPut(staticApiEndpoint, otherRequestBody, t)
+	code = DoPut(staticApiEndpoint, otherRequestBody, t)
 	if code != 204 {
 		t.Errorf("expected to be possible to update already created endpoint: expected 204 != %d", code)
 	}
